+++ src/Klab.Toolkit.Messaging.Abstractions/IMediator.cs
@@ -20,7 +20,7 @@ public interface IMediator
     /// <summary>
     /// Getter property for the local event handlers
     /// </summary>
-    ConcurrentDictionary<Type, ConcurrentBag<KeyValuePair<int, Func<EventBase, CancellationToken, Task<Result>>>>> GetLocalEventHandlers();
+    ConcurrentDictionary<Type, ConcurrentBag<Func<EventBase, CancellationToken, Task<Result>>>> GetLocalEventHandlers();
 
     /// <summary>
     /// Publishes an event and not wait until event is processed.